Port : TIFF throws ClassCastException
java/itextpdf 426639ccaf

DEV-1407 (port. DEV-1465)

+++ src/core/iTextSharp/text/pdf/codec/TIFFFaxDecoder.cs
@@ -54,7 +54,7 @@ namespace iTextSharp.text.pdf.codec {
         private int bitPointer, bytePointer;
         private byte[] data;
         private int w, h;
-        private int fillOrder;
+        private long fillOrder;
         
         // Data structures needed to store changing elements for the previous
         // and the current scanline
@@ -586,7 +586,7 @@ namespace iTextSharp.text.pdf.codec {
         * @param w
         * @param h
         */
-        public TIFFFaxDecoder(int fillOrder, int w, int h) {
+        public TIFFFaxDecoder(long fillOrder, int w, int h) {
             this.fillOrder = fillOrder;
             this.w = w;
             this.h = h;
